+++ zero/model/friend_link.model.go
@@ -18,7 +18,7 @@ type (
 		// 增删改查
 		Create(ctx context.Context, in *FriendLink) (out *FriendLink, err error)
 		Update(ctx context.Context, in *FriendLink) (out *FriendLink, err error)
-		Delete(ctx context.Context, id int) (rows int64, err error)
+		Delete(ctx context.Context, id int64) (rows int64, err error)
 		First(ctx context.Context, conditions string, args ...interface{}) (out *FriendLink, err error)
 		// 批量操作
 		BatchCreate(ctx context.Context, in ...*FriendLink) (rows int64, err error)
@@ -84,7 +84,7 @@ func (s *defaultFriendLinkModel) Update(ctx context.Context, in *FriendLink) (ou
 }
 
 // 删除FriendLink记录
-func (s *defaultFriendLinkModel) Delete(ctx context.Context, id int) (rows int64, err error) {
+func (s *defaultFriendLinkModel) Delete(ctx context.Context, id int64) (rows int64, err error) {
 	db := s.DbEngin.WithContext(ctx).Table(s.tableName)
 
 	db = db.Where("id = ?", id)
