+++ zero/model/photo.model.go
@@ -18,7 +18,7 @@ type (
 		// 增删改查
 		Create(ctx context.Context, in *Photo) (out *Photo, err error)
 		Update(ctx context.Context, in *Photo) (out *Photo, err error)
-		Delete(ctx context.Context, id int) (rows int64, err error)
+		Delete(ctx context.Context, id int64) (rows int64, err error)
 		First(ctx context.Context, conditions string, args ...interface{}) (out *Photo, err error)
 		// 批量操作
 		BatchCreate(ctx context.Context, in ...*Photo) (rows int64, err error)
@@ -85,7 +85,7 @@ func (s *defaultPhotoModel) Update(ctx context.Context, in *Photo) (out *Photo, 
 }
 
 // 删除Photo记录
-func (s *defaultPhotoModel) Delete(ctx context.Context, id int) (rows int64, err error) {
+func (s *defaultPhotoModel) Delete(ctx context.Context, id int64) (rows int64, err error) {
 	db := s.DbEngin.WithContext(ctx).Table(s.tableName)
 
 	db = db.Where("id = ?", id)
