+++ zero/model/role_api.model.go
@@ -17,7 +17,7 @@ type (
 		// 增删改查
 		Create(ctx context.Context, in *RoleApi) (out *RoleApi, err error)
 		Update(ctx context.Context, in *RoleApi) (out *RoleApi, err error)
-		Delete(ctx context.Context, id int) (rows int64, err error)
+		Delete(ctx context.Context, id int64) (rows int64, err error)
 		First(ctx context.Context, conditions string, args ...interface{}) (out *RoleApi, err error)
 		// 批量操作
 		BatchCreate(ctx context.Context, in ...*RoleApi) (rows int64, err error)
@@ -79,7 +79,7 @@ func (s *defaultRoleApiModel) Update(ctx context.Context, in *RoleApi) (out *Rol
 }
 
 // 删除RoleApi记录
-func (s *defaultRoleApiModel) Delete(ctx context.Context, id int) (rows int64, err error) {
+func (s *defaultRoleApiModel) Delete(ctx context.Context, id int64) (rows int64, err error) {
 	db := s.DbEngin.WithContext(ctx).Table(s.tableName)
 
 	db = db.Where("id = ?", id)
