+++ zero/model/role_menu.model.go
@@ -17,7 +17,7 @@ type (
 		// 增删改查
 		Create(ctx context.Context, in *RoleMenu) (out *RoleMenu, err error)
 		Update(ctx context.Context, in *RoleMenu) (out *RoleMenu, err error)
-		Delete(ctx context.Context, id int) (rows int64, err error)
+		Delete(ctx context.Context, id int64) (rows int64, err error)
 		First(ctx context.Context, conditions string, args ...interface{}) (out *RoleMenu, err error)
 		// 批量操作
 		BatchCreate(ctx context.Context, in ...*RoleMenu) (rows int64, err error)
@@ -79,7 +79,7 @@ func (s *defaultRoleMenuModel) Update(ctx context.Context, in *RoleMenu) (out *R
 }
 
 // 删除RoleMenu记录
-func (s *defaultRoleMenuModel) Delete(ctx context.Context, id int) (rows int64, err error) {
+func (s *defaultRoleMenuModel) Delete(ctx context.Context, id int64) (rows int64, err error) {
 	db := s.DbEngin.WithContext(ctx).Table(s.tableName)
 
 	db = db.Where("id = ?", id)
