+++ zero/model/user_information.model.go
@@ -18,7 +18,7 @@ type (
 		// 增删改查
 		Create(ctx context.Context, in *UserInformation) (out *UserInformation, err error)
 		Update(ctx context.Context, in *UserInformation) (out *UserInformation, err error)
-		Delete(ctx context.Context, id int) (rows int64, err error)
+		Delete(ctx context.Context, id int64) (rows int64, err error)
 		First(ctx context.Context, conditions string, args ...interface{}) (out *UserInformation, err error)
 		// 批量操作
 		BatchCreate(ctx context.Context, in ...*UserInformation) (rows int64, err error)
@@ -89,7 +89,7 @@ func (s *defaultUserInformationModel) Update(ctx context.Context, in *UserInform
 }
 
 // 删除UserInformation记录
-func (s *defaultUserInformationModel) Delete(ctx context.Context, id int) (rows int64, err error) {
+func (s *defaultUserInformationModel) Delete(ctx context.Context, id int64) (rows int64, err error) {
 	db := s.DbEngin.WithContext(ctx).Table(s.tableName)
 
 	db = db.Where("id = ?", id)
